+++ frontend/package.json
@@ -25,16 +25,13 @@
     "@radix-ui/react-separator": "^1.1.0",
     "@radix-ui/react-slot": "^1.1.0",
     "@tanstack/react-table": "^8.20.5",
-    "@types/node": "22.5.1",
-    "@types/react": "18.3.4",
-    "@types/react-dom": "18.3.0",
     "autoprefixer": "10.4.20",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cmdk": "1.0.0",
     "framer-motion": "^11.3.31",
+    "geist": "^1.3.1",
     "lucide-react": "^0.436.0",
-    "nes.css": "^2.3.0",
     "next": "~14.2.7",
     "petra-plugin-wallet-adapter": "^0.4.5",
     "postcss": "8.4.41",
@@ -47,6 +44,9 @@
     "zod": "^3.23.8"
   },
   "devDependencies": {
+    "@types/node": "22.5.1",
+    "@types/react": "18.3.4",
+    "@types/react-dom": "18.3.0",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
     "eslint": "^8.56.0",
