+++ frontend/src/app/(landing)/data/schema.ts
