+++ frontend/src/app/(landing)/data/seed.ts
