+++ frontend/tools/run.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+pnpm run dev > ./my.log 2>&1 &
+echo $! > ./dev-pid.txt
